requirements.txt
pillow==11.1.0

